pyproject.toml
[project]
name = "dr-mike-gym-tracker"
version = "0.1.0"
description = "AI-powered Telegram bot for gym workout tracking with progressive overload"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.1.1",
    "python-telegram-bot>=22.5",
    "gspread>=6.1.4",
    "google-auth>=2.35.0",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
    "openai>=2.6.1",
    "anthropic>=0.71.0",
]
